Fix logic/syntax bugs with switch statements matching multiple values
After a bit of tinkering, I figured out the cause of Issues #18, #22, and #23.  The OR (||)
conditional for multiple case values was not working as intended -- they only matched the first
value in each case instead of any of the values separated by ||.

Converting those to use the fall-through feature of the switch statement resolved the problems.